refactor: move test in service to class test
